Added detail view component and moved types to case domain
